feat(controllers): user, auth, and sales controllers created

+++ Backend/src/models/Product.js
@@ -8,7 +8,7 @@ const Product = sequelize.define(
             type: DataTypes.STRING
         }, 
         price:{ 
-            type: DataTypes.DECIMAL
+            type: DataTypes.DOUBLE
         },
         availability: {
             type: DataTypes.BOOLEAN,
